Comentarios y simplificación: - Comentada clase Point. - Comentada clase StoneFactory. Simplificada en RandomStoneFactory. Debido a este cambio, se modificó la clase GameEngine y la clase Main en todas las intancias donde se la llamaba. - Comentada la clase Size. - Comentada la clase Stone y reordenado su código.
