+++ src/main/scala/scalatetris/engine/GameEngine.scala
@@ -1,15 +1,15 @@
 package scalatetris.engine
 
-import scalatetris.environment._
+import scalatetris.environment.{Board, Size, RandomStoneFactory, Stone, Point, Statistics}
 import scalatetris.AudioManager
 import scala.util.Try
 
 //le paso por parametro el tamaño del tablero y la fabrica de piezas de tetris
-sealed class GameEngine(val boardSize: Size, val stoneFactory: StoneFactory) {
+sealed class GameEngine(val boardSize: Size) {
   private var board: Board = new Board(
     boardSize,
-    stoneFactory.createRandomStone(),
-    stoneFactory.createRandomStone()
+    RandomStoneFactory.createRandomStone(),
+    RandomStoneFactory.createRandomStone()
   )
 
   // Nivel actual del juego (0-29)
@@ -49,7 +49,7 @@ sealed class GameEngine(val boardSize: Size, val stoneFactory: StoneFactory) {
         }
       }
       
-      board = board.update(List(Stone(points)), numberOfRemovedRows, stoneFactory.createRandomStone())
+      board = board.update(List(Stone(points)), numberOfRemovedRows, RandomStoneFactory.createRandomStone())
       holdUsedThisTurn = false  // Reset del hold para la nueva pieza
       
       if (!board.isGameRunning) {
@@ -179,7 +179,7 @@ sealed class GameEngine(val boardSize: Size, val stoneFactory: StoneFactory) {
       // Actualizar el tablero solo con las piezas fijas
       val remainingStones = board.stones.tail
       board = board.updateStones(remainingStones)
-      board = board.forceNewStone(stoneFactory.createRandomStone())
+      board = board.forceNewStone(RandomStoneFactory.createRandomStone())
       holdUsedThisTurn = true
     }
     
@@ -194,8 +194,8 @@ sealed class GameEngine(val boardSize: Size, val stoneFactory: StoneFactory) {
   def restart(): Unit = {
     board = new Board(
       boardSize,
-      stoneFactory.createRandomStone(),
-      stoneFactory.createRandomStone()
+      RandomStoneFactory.createRandomStone(),
+      RandomStoneFactory.createRandomStone()
     )
     isRunning = true
     currentLevel = 0
