+++ src/main/scala/scalatetris/environment/Size.scala
@@ -1,3 +1,9 @@
 package scalatetris.environment
 
-case class Size (width: Int, height: Int)
+/**
+ * Representa el tamaño del board. Implementado como case class para permitir la comparación entre tableros y sus tamaños.
+ *
+ * @param width  Ancho del board (cantidad de columnas)
+ * @param height Alto del board (cantidad de filas)
+ */
+case class Size(width: Int, height: Int)
